requirements.txt
-e .
click==4.0
ipdb
pyyaml==3.11
requests==2.7.0
tasklib==0.9.0
